* broke out CLI sub commands into modules to increase maintainability * A small pile of linting
