+++ great_expectations/cli/cli.py
@@ -1,78 +1,27 @@
 # -*- coding: utf-8 -*-
 import click
-import json
 import logging
-import os
-import shutil
-import subprocess
-import sys
-import warnings
 
-from great_expectations.cli.init_messages import (
-    BUILD_DOCS_PROMPT,
-    COMPLETE_ONBOARDING_PROMPT,
-    GREETING,
-    LETS_BEGIN_PROMPT,
-    NEW_TEMPLATE_INSTALLED,
-    NEW_TEMPLATE_PROMPT,
-    NO_DATASOURCES_FOUND,
-    ONBOARDING_COMPLETE,
-    PROJECT_IS_COMPLETE,
-    RUN_INIT_AGAIN,
-    SLACK_LATER,
-    SLACK_SETUP_INTRO,
-    SLACK_SETUP_COMPLETE,
-    SLACK_SETUP_PROMPT,
-    SLACK_WEBHOOK_PROMPT,
-)
-from great_expectations.core import (
-    expectationSuiteSchema,
-    expectationSuiteValidationResultSchema,
-    ExpectationSuiteValidationResult,
-    NamespaceAwareExpectationSuite,
-)
-from great_expectations.datasource.generator import ManualGenerator
-from great_expectations.render.renderer.notebook_renderer import NotebookRenderer
-from great_expectations.cli.datasource import (
-    add_datasource as add_datasource_impl,
-    profile_datasource,
-    create_expectation_suite as create_expectation_suite_impl,
-    build_docs as build_documentation_impl,
-)
-from great_expectations.cli.util import cli_message, is_sane_slack_webhook
-from great_expectations.data_context import DataContext
-from great_expectations.data_asset import FileDataAsset
-from great_expectations.dataset import Dataset, PandasDataset
-import great_expectations.exceptions as ge_exceptions
+from great_expectations.cli.logging import logger, _set_up_logger
+from great_expectations.cli.datasource import datasource
+from great_expectations.cli.docs import docs
+from great_expectations.cli.init import init
+from great_expectations.cli.project import project
+from great_expectations.cli.suite import suite
+from great_expectations.cli.validation import validation
+
 from great_expectations import __version__ as ge_version
-from great_expectations import read_csv
-
-#FIXME: This prevents us from seeing a huge stack of these messages in python 2. We'll need to fix that later.
-# tests/test_cli.py::test_cli_profile_with_datasource_arg
-#   /Users/abe/Documents/superconductive/tools/great_expectations/tests/test_cli.py:294: Warning: Click detected the use of the unicode_literals __future__ import.  This is heavily discouraged because it can introduce subtle bugs in your code.  You should instead use explicit u"" literals for your unicode strings.  For more information see https://click.palletsprojects.com/python3/
-#     cli, ["profile", "my_datasource", "-d", project_root_dir])
-click.disable_unicode_literals_warning = True
-
-
-warnings.filterwarnings('ignore')
-
-logging.getLogger("great_expectations.datasource.generator.in_memory_generator").setLevel(logging.CRITICAL)
-
-try:
-    from termcolor import colored
-except ImportError:
-    colored = None
-
-
-# Take over the entire GE module logging namespace when running CLI
-logger = logging.getLogger("great_expectations")
-
 
 # TODO: consider using a specified-order supporting class for help (but wasn't working with python 2)
 @click.group()
 @click.version_option(version=ge_version)
-@click.option('--verbose', '-v', is_flag=True, default=False,
-              help='Set great_expectations to use verbose output.')
+@click.option(
+    "--verbose",
+    "-v",
+    is_flag=True,
+    default=False,
+    help="Set great_expectations to use verbose output.",
+)
 def cli(verbose):
     """great_expectations command-line interface"""
     if verbose:
@@ -82,709 +31,17 @@ def cli(verbose):
         logger.setLevel(logging.DEBUG)
 
 
-@cli.group()
-def project():
-    """project operations"""
-    pass
-
-
-@cli.group()
-def datasource():
-    """datasource operations"""
-    pass
-
-
-@cli.group()
-def suite():
-    """expectation suite operations"""
-    pass
-
-
-@cli.group()
-def docs():
-    """data docs operations"""
-    pass
-
-
-@cli.group()
-def validation():
-    """validation operations"""
-    pass
-
-
-@validation.command(name="csv")
-@click.argument('dataset')
-@click.argument('expectation_suite_file')
-@click.option('--evaluation_parameters', '-p', default=None,
-              help='Path to a file containing JSON object used to evaluate parameters in expectations config.')
-@click.option('--result_format', '-o', default="SUMMARY",
-              help='Result format to use when building evaluation responses.')
-@click.option('--catch_exceptions', '-e', default=True, type=bool,
-              help='Specify whether to catch exceptions raised during evaluation of expectations (defaults to True).')
-@click.option('--only_return_failures', '-f', default=False, type=bool,
-              help='Specify whether to only return expectations that are not met during evaluation '
-                   '(defaults to False).')
-@click.option('--custom_dataset_module', '-m', default=None,
-              help='Path to a python module containing a custom dataset class.')
-@click.option('--custom_dataset_class', '-c', default=None,
-              help='Name of the custom dataset class to use during evaluation.')
-def validation_csv(
-        dataset,
-        expectation_suite_file,
-        evaluation_parameters,
-        result_format,
-        catch_exceptions, only_return_failures, custom_dataset_module, custom_dataset_class):
-    """Validate a CSV file against an expectation suite.
-
-    DATASET: Path to a file containing a CSV file to validate using the provided expectation_suite_file.
-
-    EXPECTATION_SUITE_FILE: Path to a file containing a valid great_expectations expectations suite to use to \
-validate the data.
-    """
-
-    """
-    Read a dataset file and validate it using an expectation suite saved in another file. Uses parameters defined in 
-    the dispatch method.
-
-    :param parsed_args: A Namespace object containing parsed arguments from the dispatch method.
-    :return: The number of unsuccessful expectations
-    """
-    expectation_suite_file = expectation_suite_file
-    with open(expectation_suite_file, 'r') as infile:
-        expectation_suite = expectationSuiteSchema.load(json.load(infile)).data
-
-    if evaluation_parameters is not None:
-        evaluation_parameters = json.load(
-            open(evaluation_parameters, "r"))
-
-    # Use a custom data_asset module and class if provided. Otherwise infer from the expectation suite
-    if custom_dataset_module:
-        sys.path.insert(0, os.path.dirname(
-            custom_dataset_module))
-        module_name = os.path.basename(
-            custom_dataset_module).split('.')[0]
-        custom_module = __import__(str(module_name))
-        dataset_class = getattr(
-            custom_module, custom_dataset_class)
-    elif expectation_suite.data_asset_type is not None:
-        if (expectation_suite.data_asset_type == "Dataset" or
-                expectation_suite.data_asset_type == "PandasDataset"):
-            dataset_class = PandasDataset
-        elif expectation_suite.data_asset_type.endswith("Dataset"):
-            logger.info("Using PandasDataset to validate dataset of type %s." %
-                        expectation_suite.data_asset_type)
-            dataset_class = PandasDataset
-        elif expectation_suite.data_asset_type == "FileDataAsset":
-            dataset_class = FileDataAsset
-        else:
-            cli_message(
-                "Unrecognized data_asset_type %s. You may need to specify "
-                "custom_dataset_module and custom_dataset_class." % expectation_suite.data_asset_type
-            )
-            return -1
-    else:
-        dataset_class = PandasDataset
-
-    if issubclass(dataset_class, Dataset):
-        da = read_csv(dataset, expectation_suite=expectation_suite,
-                      dataset_class=dataset_class)
-    else:
-        da = dataset_class(dataset, config=expectation_suite)
-
-    result = da.validate(
-        evaluation_parameters=evaluation_parameters,
-        result_format=result_format,
-        catch_exceptions=catch_exceptions,
-        only_return_failures=only_return_failures,
-    )
-
-    # Note: Should this be rendered through cli_message?
-    # Probably not, on the off chance that the JSON object contains <color> tags
-    print(json.dumps(expectationSuiteValidationResultSchema.dump(result).data, indent=2))
-    sys.exit(result['statistics']['unsuccessful_expectations'])
-
-
-@cli.command()
-@click.option(
-    '--target_directory',
-    '-d',
-    default="./",
-    help='The root of the project directory where you want to initialize Great Expectations.'
-)
-@click.option(
-    # Note this --no-view option is mostly here for tests
-    "--view/--no-view",
-    help="By default open in browser unless you specify the --no-view flag",
-    default=True
-)
-def init(target_directory, view):
-    """
-    Initialize a new Great Expectations project.
-
-    This guided input walks the user through setting up a new project and also
-    onboards a new developer in an existing project.
-
-    It scaffolds directories, sets up notebooks, creates a project file, and
-    appends to a `.gitignore` file.
-    """
-    target_directory = os.path.abspath(target_directory)
-    ge_dir = _get_full_path_to_ge_dir(target_directory)
-    ge_yml = os.path.join(ge_dir, DataContext.GE_YML)
-
-    cli_message(GREETING)
-
-    # TODO this should be a property
-    if os.path.isfile(ge_yml):
-        if DataContext.all_uncommitted_directories_exist(ge_dir) and \
-                DataContext.config_variables_yml_exist(ge_dir):
-            # Ensure the context can be instantiated
-            try:
-                _ = DataContext(ge_dir)
-                cli_message(PROJECT_IS_COMPLETE)
-            except ge_exceptions.DataContextError as e:
-                cli_message("<red>{}</red>".format(e))
-                exit(5)
-        else:
-            _complete_onboarding(target_directory)
-
-        try:
-            # if expectations exist, offer to build docs
-            context = DataContext(ge_dir)
-            if context.list_expectation_suite_keys():
-                if click.confirm(BUILD_DOCS_PROMPT, default=True):
-                    context.build_data_docs()
-                    context.open_data_docs()
-        except ge_exceptions.DataContextError as e:
-            cli_message("<red>{}</red>".format(e))
-    else:
-        if not click.confirm(LETS_BEGIN_PROMPT, default=True):
-            cli_message(RUN_INIT_AGAIN)
-            exit(0)
-        try:
-            context, datasource_name, data_source_type = _create_new_project(target_directory)
-            if not datasource_name:  # no datasource was created
-                return
-
-            # we need only one of the values returned here - profiling_results
-            datasource_name, generator_name, data_asset_name, batch_kwargs, profiling_results = \
-                create_expectation_suite_impl(
-                        context,
-                        datasource_name=datasource_name,
-                        show_intro_message=False,
-                        additional_batch_kwargs={"limit": 1000},
-                        open_docs=view)
-
-            notebook_renderer = NotebookRenderer()
-
-            for result in profiling_results["results"]:
-                # TODO brittle
-                suite = result[0]
-                assert isinstance(suite, NamespaceAwareExpectationSuite)
-                suite_name = suite.expectation_suite_name
-
-                validation_result = result[1]
-                assert isinstance(validation_result, ExpectationSuiteValidationResult)
-                batch_kwargs = validation_result.meta.get("batch_kwargs")
-                data_asset_identifier = validation_result.meta.get("data_asset_name")
-                human_data_asset_name = data_asset_identifier.generator_asset
-
-                logger.debug(f"\nRendering a notebook for {human_data_asset_name} and suite {suite_name}")
-                logger.debug(f"batch_kwargs: {batch_kwargs}")
-                logger.debug(f"datasource_name: {datasource_name}")
-
-                notebook_name = f"{human_data_asset_name}_{suite_name}.ipynb"
-                notebook_path = os.path.join(context.root_directory, "notebooks", notebook_name)
-                notebook_renderer.render_to_disk(suite, batch_kwargs, notebook_path)
-
-
-            # TODO maybe loop over profiling results since they contain suites and validation results
-            # for data_asset in data_assets:
-            #     # TODO brittle
-            #     # print("\n\n\n")
-            #     # print(profiling_results)
-            #     # print("\n\n\n")
-            #     # sys.exit(1)
-            #     # batch_kwargs = profiling_results["results"][0][1]["meta"]["batch_kwargs"]
-            #
-            #     print(f"data_asset: {data_asset}")
-            #     suite = context.get_expectation_suite(
-            #         data_asset_name=data_asset,
-            #         expectation_suite_name=str(profiler.__name__)
-            #     )
-            #     suite_name = suite.expectation_suite_name
-            #
-            #     print(f"\nRendering a notebook for {data_asset} and suite {suite_name}")
-            #
-            #     # batch_kwargs = context.build_batch_kwargs(data_asset, partition_id="profiler")
-            #     batch_kwargs = context.yield_batch_kwargs(data_asset)
-            #     batch_kwargs_by_data_asset[data_asset] = batch_kwargs
-            #     print(f"batch_kwargs: {batch_kwargs}")
-            #     print(f"datasource_name: {datasource_name}")
-            #
-            #     notebook_name = f"{data_asset}_{suite_name}.ipynb"
-            #     notebook_renderer.render_to_disk(suite, batch_kwargs, os.path.join(context.root_directory, notebook_name))
-                cli_message("""\n<cyan>Great Expectations is now set up.</cyan>""")
-        except ge_exceptions.DataContextError as e:
-            cli_message("<red>{}</red>".format(e))
-
-
-
-def _slack_setup(context):
-    webhook_url = None
-    cli_message(SLACK_SETUP_INTRO)
-    if not click.confirm(SLACK_SETUP_PROMPT, default=True):
-        cli_message(SLACK_LATER)
-        return context
-    else:
-        webhook_url = click.prompt(SLACK_WEBHOOK_PROMPT, default="")
-
-    while not is_sane_slack_webhook(webhook_url):
-        cli_message("That URL was not valid.\n")
-        if not click.confirm(SLACK_SETUP_PROMPT, default=True):
-            cli_message(SLACK_LATER)
-            return context
-        webhook_url = click.prompt(SLACK_WEBHOOK_PROMPT, default="")
-
-    context.save_config_variable("validation_notification_slack_webhook", webhook_url)
-    cli_message(SLACK_SETUP_COMPLETE)
-
-    return context
-
-
-def _get_full_path_to_ge_dir(target_directory):
-    return os.path.abspath(os.path.join(target_directory, DataContext.GE_DIR))
-
-
-def _create_new_project(target_directory):
-    try:
-        context = DataContext.create(target_directory)
-        datasource_name, data_source_type = add_datasource_impl(context)
-        return context, datasource_name, data_source_type
-    except ge_exceptions.DataContextError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        sys.exit(-1)
-
-
-def _complete_onboarding(target_dir):
-    if click.confirm(COMPLETE_ONBOARDING_PROMPT, default=True):
-        DataContext.create(target_dir)
-        cli_message(ONBOARDING_COMPLETE)
-    else:
-        cli_message(RUN_INIT_AGAIN)
-
-
-@datasource.command(name="new")
-@click.option(
-    '--directory',
-    '-d',
-    default=None,
-    help="The project's great_expectations directory."
-)
-@click.option(
-    "--view/--no-view",
-    help="By default open in browser unless you specify the --no-view flag",
-    default=True
-)
-def datasource_new(directory, view):
-    """Add a new datasource to the data context."""
-    try:
-        context = DataContext(directory)
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        return
-    except ge_exceptions.ZeroDotSevenConfigVersionError as err:
-        _offer_to_install_new_template(err, context.root_directory)
-
-    datasource_name, data_source_type = add_datasource_impl(context)
-
-    if not datasource_name:  # no datasource was created
-        return
-
-    # TODO do we really want to "profile" every new datasource?
-    profile_datasource(context, datasource_name, open_docs=view)
-
-
-@datasource.command(name="list")
-@click.option(
-    '--directory',
-    '-d',
-    default=None,
-    help="The project's great_expectations directory."
-)
-def datasource_list(directory):
-    """List known datasources."""
-    try:
-        context = DataContext(directory)
-        datasources = context.list_datasources()
-        # TODO Pretty up this console output
-        cli_message(str([d for d in datasources]))
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        return
-    except ge_exceptions.ZeroDotSevenConfigVersionError as err:
-        _offer_to_install_new_template(err, context.root_directory)
-
-
-@datasource.command(name="profile")
-@click.argument('datasource_name', default=None, required=False)
-@click.option('--data_assets', '-l', default=None,
-              help='Comma-separated list of the names of data assets that should be profiled. Requires datasource_name specified.')
-@click.option('--profile_all_data_assets', '-A', is_flag=True, default=False,
-              help='Profile ALL data assets within the target data source. '
-                   'If True, this will override --max_data_assets.')
-@click.option(
-    "--directory",
-    "-d",
-    default=None,
-    help="The project's great_expectations directory."
-)
-@click.option('--batch_kwargs', default=None,
-              help='Additional keyword arguments to be provided to get_batch when loading the data asset. Must be a valid JSON dictionary')
-@click.option(
-    "--view/--no-view",
-    help="By default open in browser unless you specify the --no-view flag",
-    default=True
-)
-def datasource_profile(datasource_name, data_assets, profile_all_data_assets, directory, view, batch_kwargs):
-    """
-    Profile a datasource
-
-    If the optional data_assets and profile_all_data_assets arguments are not specified, the profiler will check
-    if the number of data assets in the datasource exceeds the internally defined limit. If it does, it will
-    prompt the user to either specify the list of data assets to profile or to profile all.
-    If the limit is not exceeded, the profiler will profile all data assets in the datasource.
-
-    :param datasource_name: name of the datasource to profile
-    :param data_assets: if this comma-separated list of data asset names is provided, only the specified data assets will be profiled
-    :param profile_all_data_assets: if provided, all data assets will be profiled
-    :param directory:
-    :param view: Open the docs in a browser
-    :param batch_kwargs: Additional keyword arguments to be provided to get_batch when loading the data asset.
-    :return:
-    """
-
-    try:
-        context = DataContext(directory)
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        return
-    except ge_exceptions.ZeroDotSevenConfigVersionError as err:
-        _offer_to_install_new_template(err, context.root_directory)
-        return
-
-    if batch_kwargs is not None:
-        batch_kwargs = json.loads(batch_kwargs)
-
-    if datasource_name is None:
-        datasources = [datasource["name"] for datasource in context.list_datasources()]
-        if not datasources:
-            cli_message(NO_DATASOURCES_FOUND)
-            sys.exit(-1)
-        elif len(datasources) > 1:
-            cli_message(
-                "<red>Error: please specify the datasource to profile. "\
-                "Available datasources: " + ", ".join(datasources) + "</red>"
-            )
-            sys.exit(-1)
-        else:
-            profile_datasource(
-                context,
-                datasources[0],
-                data_assets=data_assets,
-                profile_all_data_assets=profile_all_data_assets,
-                open_docs=view,
-                additional_batch_kwargs=batch_kwargs
-            )
-    else:
-        profile_datasource(
-            context,
-            datasource_name,
-            data_assets=data_assets,
-            profile_all_data_assets=profile_all_data_assets,
-            open_docs=view,
-            additional_batch_kwargs=batch_kwargs
-        )
-
-
-@suite.command(name="edit")
-@click.argument("data_asset")
-@click.argument("suite")
-@click.option(
-    '--batch-kwargs',
-    default=None,
-    help='Additional keyword arguments to be provided to get_batch when loading \
-the data asset. Must be a valid JSON dictionary'
-)
-@click.option(
-    "--directory",
-    "-d",
-    default=None,
-    help="The project's great_expectations directory.",
-)
-@click.option(
-    "--jupyter/--no-jupyter",
-    is_flag=True,
-    help="By default launch jupyter notebooks unless you specify the --no-jupyter flag",
-    default=True,
-)
-def suite_edit(
-    data_asset, suite, directory, jupyter, batch_kwargs
-):
-    """Edit an existing suite with a jupyter notebook."""
-    try:
-        context = DataContext(directory)
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        return
-    except ge_exceptions.ZeroDotSevenConfigVersionError as err:
-        _offer_to_install_new_template(err, context.root_directory)
-        return
-
-    try:
-        normalized_data_asset_name = context.normalize_data_asset_name(data_asset)
-    except ge_exceptions.AmbiguousDataAssetNameError as e:
-        cli_message(
-            """<cyan>Your data_asset name of {} was not specific enough.
-Please re-run with one of these selected data assets:
-  - '{}""".format(data_asset,
-                "'\n  - '".join([str(ds) for ds in e.candidates]))
-            + "'</cyan>"
-        )
-        sys.exit(-1)
-    except ge_exceptions.DataContextError as e:
-        # TODO consider increasing precision of other DataContextErrors to improve usability
-        # TODO Add suggestion to user to run `ge list-data-assets`
-        cli_message("<red>{}</red>".format(e.message))
-        sys.exit(-1)
-
-    suite = suite.rstrip(".json")
-    suite = _load_suite(context, normalized_data_asset_name, suite)
-
-    data_source = context.get_datasource(normalized_data_asset_name.datasource)
-    generator = data_source.get_generator(normalized_data_asset_name.generator)
-
-    if batch_kwargs:
-        batch_kwargs = json.loads(batch_kwargs)
-    # elif suite.get_original_batch_kwargs():
-        # TODO this functionality doesn't actually exist yet
-        # batch_kwargs = suite.get_original_batch_kwargs()
-    elif isinstance(generator, ManualGenerator):
-        generator_asset = suite.data_asset_name.generator_asset
-        batch_kwargs = generator.yield_batch_kwargs(generator_asset)
-    else:
-        batch_kwargs = context.yield_batch_kwargs(suite.data_asset_name)
-
-    if not batch_kwargs:
-        cli_message(
-            "<red>Attempting to use a configured generator to build batch_"
-            "kwargs. You may need to review the generator configuration to "
-            "ensure you can get the desired batch.</red>"
-        )
-
-    human_data_asset_name = suite.data_asset_name.generator_asset
-    notebook_name = f"{human_data_asset_name}_{suite.expectation_suite_name}.ipynb"
-
-    notebook_path = os.path.join(context.GE_EDIT_NOTEBOOK_DIR, notebook_name)
-    NotebookRenderer().render_to_disk(suite, batch_kwargs, notebook_path)
-
-    cli_message(
-        "To continue editing this suite, run <green>jupyter notebook {}</green>".format(
-            notebook_path
-        )
-    )
-
-    if jupyter:
-        subprocess.call(["jupyter", "notebook", notebook_path])
-
-
-def _load_suite(context, data_asset_name, suite_name):
-    try:
-        suite = context.get_expectation_suite(data_asset_name, suite_name)
-    except ge_exceptions.DataContextError as e:
-        cli_message(
-            "<red>Could not locate a suite named {} for {}</red>".format(
-                suite_name, data_asset_name)
-        )
-        logger.info(e)
-        sys.exit(-1)
-    return suite
-
-
-@docs.command(name="build")
-@click.option(
-    '--directory',
-    '-d',
-    default=None,
-    help="The project's great_expectations directory."
-)
-@click.option('--site_name', '-s',
-              help='The site for which to generate documentation. See data_docs section in great_expectations.yml')
-@click.option(
-    "--view/--no-view",
-    help="By default open in browser unless you specify the --no-view flag",
-    default=True
-)
-def docs_build(directory, site_name, view=True):
-    """Build Data Docs for a project."""
-    try:
-        context = DataContext(directory)
-        build_documentation_impl(
-            context,
-            site_name=site_name
-        )
-        if view:
-            context.open_data_docs()
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        sys.exit(1)
-    except ge_exceptions.ZeroDotSevenConfigVersionError as err:
-        _offer_to_install_new_template(err, context.root_directory)
-        return
-    except ge_exceptions.PluginModuleNotFoundError as err:
-        cli_message(err.cli_colored_message)
-        sys.exit(1)
-    except ge_exceptions.PluginClassNotFoundError as err:
-        cli_message(err.cli_colored_message)
-        sys.exit(1)
-
-
-@project.command(name="check-config")
-@click.option(
-    '--directory',
-    '-d',
-    default="./great_expectations",
-    help="The project's great_expectations directory."
-)
-def project_check_config(directory):
-    """Check a config for validity and help with migrations."""
-    cli_message("Checking your config files for validity...\n")
-
-    try:
-        is_config_ok, error_message = do_config_check(directory)
-        if is_config_ok:
-            cli_message("<green>Your config file appears valid!</green>")
-        else:
-            cli_message("Unfortunately, your config appears to be invalid:\n")
-            cli_message("<red>{}</red>".format(error_message))
-            sys.exit(1)
-    except ge_exceptions.ZeroDotSevenConfigVersionError as err:
-        _offer_to_install_new_template(err, directory)
-
-
-@suite.command(name="new")
-@click.option('--data_asset', '-da', default=None,
-              help='Fully qualified data asset name (datasource/generator/generator_asset)')
-@click.option('--suite', '-es', default=None,
-              help='Expectation suite name.')
-@click.option(
-    "--directory",
-    "-d",
-    default=None,
-    help="The project's great_expectations directory."
-)
-@click.option('--batch_kwargs', default=None,
-              help='Additional keyword arguments to be provided to get_batch when loading the data asset. Must be a valid JSON dictionary')
-def suite_new(data_asset, suite, directory, batch_kwargs):
-    """
-    Create a new expectation suite.
-
-    Great Expectations will choose a couple of columns and generate expectations about them
-    to demonstrate some examples of assertions you can make about your data.
-    """
-    try:
-        context = DataContext(directory)
-    except ge_exceptions.ConfigNotFoundError as err:
-        cli_message("<red>{}</red>".format(err.message))
-        return
-    except ge_exceptions.ZeroDotSevenConfigVersionError as err:
-        _offer_to_install_new_template(err, context.root_directory)
-        return
-
-    if batch_kwargs is not None:
-        batch_kwargs = json.loads(batch_kwargs)
-
-    datasource_name = None
-    generator_name = None
-    generator_asset = None
-    if data_asset is not None:
-        try:
-            data_asset_id = context.normalize_data_asset_name(data_asset)
-            datasource_name = data_asset_id.datasource
-            generator_name = data_asset_id.generator
-            generator_asset = data_asset_id.generator_asset
-        except ge_exceptions.AmbiguousDataAssetNameError as e:
-            cli_message(
-                """<cyan>Your data_asset name of {} was not specific enough.
-    No worries - we will help you to specify the data asset name
-      - '{}""".format(data_asset)
-                + "'</cyan>"
-            )
-
-    create_expectation_suite_impl(
-        context,
-        datasource_name=datasource_name,
-        generator_name=generator_name,
-        generator_asset=generator_asset,
-        batch_kwargs=batch_kwargs,
-        expectation_suite_name=suite,
-        additional_batch_kwargs=None,
-        show_intro_message=False,
-        open_docs=True)
-
-
-def _offer_to_install_new_template(err, ge_dir):
-    ge_dir = os.path.abspath(ge_dir)
-    cli_message("<red>{}</red>".format(err.message))
-    ge_yml = os.path.join(ge_dir, DataContext.GE_YML)
-    archived_yml = ge_yml + ".archive"
-
-    if click.confirm(
-        NEW_TEMPLATE_PROMPT.format(ge_yml, archived_yml),
-        default=True
-    ):
-        # archive existing project config
-        shutil.move(ge_yml, archived_yml)
-        DataContext.write_project_template_to_disk(ge_dir)
-
-        cli_message(NEW_TEMPLATE_INSTALLED.format("file://" + ge_yml, "file://" + archived_yml))
-    else:
-        cli_message(
-            """\nOK. To continue, you will need to upgrade your config file to the latest format.
-  - Please see the docs here: <blue>https://docs.greatexpectations.io/en/latest/reference/data_context_reference.html</blue>
-  - We are super sorry about this breaking change! :]
-  - If you are running into any problems, please reach out on Slack and we can
-    help you in realtime: https://greatexpectations.io/slack"""
-        )
-    sys.exit(0)
-
-
-def do_config_check(target_directory):
-    try:
-        DataContext(context_root_dir=target_directory)
-        return True, None
-    except (
-            ge_exceptions.InvalidConfigurationYamlError,
-            ge_exceptions.InvalidTopLevelConfigKeyError,
-            ge_exceptions.MissingTopLevelConfigKeyError,
-            ge_exceptions.InvalidConfigValueTypeError,
-            ge_exceptions.UnsupportedConfigVersionError,
-            ge_exceptions.DataContextError,
-            ge_exceptions.PluginClassNotFoundError
-            ) as err:
-        return False, err.message
-
-
-def _set_up_logger():
-    handler = logging.StreamHandler()
-    formatter = logging.Formatter('%(message)s')
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
-    logger.setLevel(logging.INFO)
+cli.add_command(datasource)
+cli.add_command(docs)
+cli.add_command(init)
+cli.add_command(project)
+cli.add_command(suite)
+cli.add_command(validation)
 
 
 def main():
     cli()
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
